backend/requirements.txt
Flask==2.0.1Flask

Pillow==9.5.0Pillow

torch>=1.7.0flask-cors

torchvision>=0.8.1google-generativeai

numpy>=1.18.5python-dotenv

opencv-python>=4.1.2requests

PyYAML>=5.3.1
requests>=2.23.0
tqdm>=4.41.0
matplotlib>=3.2.2
seaborn>=0.11.0
pandas>=1.1.4
flask-cors==4.0.0
ultralytics==8.0.196